Simplify determining resource collection namespace. Moved logic in ResourceCollection mocks to trait
