+++ src/Service/ResourceCollection.php
@@ -12,21 +12,12 @@ abstract class ResourceCollection extends Resource
     protected $resources = [];
 
     /**
-     * @var string
-     */
-    private $namespace;
-
-    abstract protected function getNamespaceSuffix();
-
-    /**
      * Return the namespace of Service class names
      * @return string
      */
     protected function getNamespace()
     {
-        $parts = explode('\\', get_called_class(), -1);
-        $parts[] = $this->getNamespaceSuffix();
-        return '\\' . implode('\\', $parts);
+        return get_called_class();
     }
 
     /**
@@ -42,10 +33,7 @@ abstract class ResourceCollection extends Resource
     public function __get($name)
     {
         if (!isset($this->resources[$name])) {
-            if (!$this->namespace) {
-                $this->namespace = $this->getNamespace();
-            }
-            $clsName = $this->namespace . '\\' . ucfirst($name);
+            $clsName = $this->getNamespace() . '\\' . ucfirst($name);
             if (!is_subclass_of($clsName, '\Nexmo\Service\Resource')) {
                 throw new Exception("Class $clsName is not a Nexmo Resource");
             }
